Merge: Menu 관련 MapStruct 라이브러리 적용 코드삭제 , Dto 경로수정, 기타 코드 리팩터링  (#92)
* Feat: 엑세스토큰 재발급 사용자/점주 엔드포인트 통일 (프론트 요청사항)

- api 한개로 통합
- RefreshTokenService 내부에서 USER_ROLE을 통해 회원타입을 분기하고 각각의 레포지토리를 조회해 User/Restaurant엔티티를 받아 새로운 토큰 발급

* Refactor: Swagger Api 태그 수정

* Feat: 웹소켓 관련 의존성 추가

* Feat: 웹소켓으로 총 대기인원, 고객의 대기인원 업데이트 하는 기능 추가

고객이 큐 등록 시 -> 식당에 전체 큐 인원 업데이트
식당이 큐 당길 경우 -> 각 고객에게 업데이트 된 본인의 대기번호 업데이트

* Feat: Queue 관련 예외 포함 정의되지 않은 예외 추가

* Test: 웹소켓 업데이트 테스트코드 작성

* Test: 웹소켓 업데이트 테스트코드 작성

* Feat: 대기열이 비어있는 경우 예외케이스 처리 로직 추가

* Test: 대기열(큐) 서비스 관련 테스트코드 작성

* Feat: Queue 웹소켓 코드 삭제, 예외처리 추가

- 웹소켓을 사용해 클라이언트에 웨이팅번호를 업데이트 하는
  코드를 삭제했습니다.(프론트팀에서 삭제 요청)
- 웨이팅 중복등록 , 등록되지 않은 웨이팅 조회 등의 예외처리를 추가하였습니다.

* Feat: Queue 관련 예외 추가

Queue관련 예외 메세지를 추가하고, Slf4j 를 사용하여
에러 메세지를 서버에도 log를 남기도록 수정했습니다.

* Test: QueueService 코드 추가 및 수정에 따른 테스트코드 수정

웹소켓제거, 예외처리 등에 따라 변경된 내용을 테스트 코드에 적용했습니다.

* Refactor: 디렉토리 변경으로 인한 import문 수정

* Refactor: Conflict 처리 하며 누락된 import문 추가

* Comment: QueueService 내 메서드별 주석 수정 및 보완

* Refactor: 조회 메서드에 Transactional(readonly=true) 어노테이션추가

* Test: 테스트코드 DisplayName 회원 -> 사용자로 용어통일

* Refactor: api경로 변경 및 조회 메서드에 Transactional(readonly=true) 어노테이션추가

* Refactor: QueueService 스트림을 사용한 코드 리팩터링

UserServce:
데이터 양이 작고 연산이 간단하므로 병렬 처리를 사용하지 않는 한
성능 저하가 크게 발생하지 않을 것으로 판단되어
스트림을 사용하였음.(코드의 가독이 향상,반복문이나 조건문의 중첩 제거)

GlobalExceptionHandler:
log.error(exception.getMessage(),exception) 을 사용하여
예외메세지뿐아니라 예외발생위치 등을 로깅하도록 처리

* Refactor: 불필요한 어노테이션 제거

* Refactor: MapStruct 라이브러리 적용 코드삭제 , Dto 경로수정, 기타 코드 리팩터링

- MapStruct를 사용한 코드를삭제하고 .of() 메서드를 만들어 적용했습니다.
- Dto의 경로를 수정했습니다.
- 단위테스트/api테스트를통해 버그들을 확인 후 수정하였습니다.
  - 비밀번호 초기화 시 임시비밀번호에 "-" 포함되던 버그 수정
  - 회원정보 수정 시 휴대폰번호가 nickname 컬럼에 저장되던 버그 수정

* Test: MapStruct 코드제거 및 Dto경로 변경에 의한 테스트코드수정

- 이외에도 AssertJ 를 사용하지 않은 부분 수정했습니다.

* Test: 불필요한 어노테이션 제거

* Refactor: MapStruct 코드 삭제 및 Dto 경로변경으로 인한 import문 수정

* Feat: 컨트롤러 내 반환 값 수정

* Refactor: Kakao 로그인 관련 Dto 경로 수정

* Refactor: Queue 관련 MapStruct 라이브러리 적용 코드삭제 , Dto 경로수정, 기타 코드 리팩터링

- MapStruct를 사용한 코드를삭제하고 .of() 메서드를 만들어 적용했습니다.
- Dto의 경로를 수정했습니다.

* Test: MapStruct 코드제거 및 Dto경로 변경에 의한 테스트코드수정

* Refactor: Menu 관련 MapStruct 라이브러리 적용 코드삭제 , Dto 경로수정, 기타 코드 리팩터링

- MapStruct를 사용한 코드를삭제하고 .of() 메서드를 만들어 적용했습니다.
- Dto의 경로를 수정했습니다.

* Test: MapStruct 코드제거 및 Dto경로 변경에 의한 테스트코드수정

* Fix: MenuDto 의 of 메서드에 id값 추가가 누락되어 발생한 버그 수정
