+++ src/main/java/com/bttf/queosk/controller/MenuController.java
@@ -1,10 +1,10 @@
 package com.bttf.queosk.controller;
 
 import com.bttf.queosk.config.JwtTokenProvider;
-import com.bttf.queosk.dto.menudto.MenuCreationForm;
-import com.bttf.queosk.dto.menudto.MenuDto;
-import com.bttf.queosk.dto.menudto.MenuStatusForm;
-import com.bttf.queosk.dto.menudto.MenuUpdateForm;
+import com.bttf.queosk.dto.MenuCreationForm;
+import com.bttf.queosk.dto.MenuDto;
+import com.bttf.queosk.dto.MenuStatusForm;
+import com.bttf.queosk.dto.MenuUpdateForm;
 import com.bttf.queosk.service.ImageService;
 import com.bttf.queosk.service.MenuService;
 import io.swagger.annotations.Api;
