+++ src/main/java/com/bttf/queosk/dto/QueueResponseForRestaurant.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.dto.queuedto;
+package com.bttf.queosk.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -14,4 +14,11 @@ import java.util.List;
 public class QueueResponseForRestaurant {
     private Integer totalQueue;
     private List<QueueDto> queueDtoList;
+
+    public static QueueResponseForRestaurant of(List<QueueDto> queueDtos) {
+        return QueueResponseForRestaurant.builder()
+                .totalQueue(queueDtos.size())
+                .queueDtoList(queueDtos)
+                .build();
+    }
 }
