--- src/main/java/com/bttf/queosk/dto/menudto/MenuStatusForm.java
@@ -1,15 +1,0 @@
-package com.bttf.queosk.dto.menudto;
-
-import com.bttf.queosk.enumerate.MenuStatus;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MenuStatusForm {
-    private MenuStatus status;
-}
