--- src/main/java/com/bttf/queosk/dto/menudto/MenuUpdateForm.java
@@ -1,15 +1,0 @@
-package com.bttf.queosk.dto.menudto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MenuUpdateForm {
-    private String name;
-    private Long price;
-}
